update css style and file structure
