--- src/pages/MyReports/components/Content/components/ReportList/ReportList.tsx
@@ -1,6 +1,0 @@
-import { FC } from 'react'
-import ListItems from './components/ListItems'
-
-const ReportList: FC = () => <ListItems />
-
-export default ReportList
