--- src/pages/MyReports/components/Content/components/ReportList/components/ListItems/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ListItems'
