--- src/pages/MyReports/components/Content/components/ReportList/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ReportList'
