+++ src/pages/MyReports/components/Content/components/ReportViewToggle/components/ListView/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ListView'
